Initial Commit
I had previously been working on a clone of the ExoPlayer project from
GitHub, but pushing that to GitHub as a new private project meant that
all collaborators on ExoPlayer could also see and access the new
project.

Since we’re working on unpublished algorithms, we don’t want this (no
unpublished algorithms were part of the project at the time it was
accidentally shared).

Here are the last few commit messages from the old project (oldest
first):

Add MISL Video to Demo Application:

* I’ve included the MISL video in the JSON file.

Import Algorithm Classes:

* AlgorithmTrackSelection and its implementation
DefaultAlgorithmTrackSelection are intended to delegate to an
implementation of AdaptationAlgorithm to perform the actual algorithm.

* The BasicAdaptationAlgorithm at the moment just uses a BandwidthMeter
to get an estimate of the bitrate, and decides based on that.

Add MISL Module and Move Classes to It:

*
The MISL module is a nice way to keep the classes that have been added
organised.

* The demo module now has a dependency on this module, and so can
access all the new classes, and the MISL module has a dependency on the
ExoPlayer libraries, and so can access all those classes (that are
public).

Replace ChunkSampleStream and DashMediaPeriod with Interfaces:

* For each, I have replaced them with an interface and a Default class
implementing that interface (e.g. DefaultChunkSampleStream). I have
also changed DashMediaPeriod from package-private visibility to public
visibility.

* These changes will facilitate replacement of these classes with
custom ones.

Add getID() to MISLDashMediaPeriod:

* When I added this method to the DashMediaPeriod interface, I forgot
to implement it in this class.

* This class is not really implemented yet anyway, but will be soon.