Copy DefaultDashMediaPeriod into MISLDashMediaPeriod
I had thought that I’d be able to change what I needed either by
composing an instance of DefaultDashMediaPeriod or by extending it, but
either way I’ll have to duplicate a lot of the implementation.

Instead, I’ve just copied the code of the existing class, and I’m going
to make small changes. I think this is the easiest way.