Make AlgorithmTrackSelection Implement AlgorithmInfoProvider
For the moment, it makes sense for the `AlgorithmTrackSelection` to be
the info provider, and it can pass itself to the `AdaptationAlgorithms`
when it’s instantiated.

The stall duration data isn’t accessible from `AlgorithmTrackSelection`
yet though, and I think that may change whether it makes sense for it
to be the info provider.