Remove `currentBufferLevelUs` from `AlgorithmInfoProvider`
I think it makes more sense to continue letting the buffer level be
passed into the track selection and from there to the
`AdaptationAlgorithm`.

Moving it to `AlgorithmInfoProvider` is nice in one sense because then
all the info is in the same place, but it’s doing more work with no
real benefit that I can see, so I’m taking it out.