Remove BandwidthMeter from DefaultAlgorithmTrackSelection
I’ve changed things so that each AdaptationAlgorithm will have its own
listener, and those are managed independently of the
DefaultAlgorithmTrackSelection – that way I can vary how that works for
different algorithms.

I replaced the BandwidthMeter with an AdaptationAlgorithm.Factory
parameter so that the algorithm can be selected from the code using the
class. Factory.createAlgorithm() is then called in the constructor
after calling the super constructor, so that `tracks` will be in sorted
order (sorting is done by BaseTrackSelection).