Use BasicAdaptationAlgorithm with an AlgorithmListener
I’ve replaced the BandwidthMeter with an AlgorithmListener, and the
code now creates a BasicAdaptationAlgorithmFactory which it passes to a
DefaultAlgorithmTrackSelection.

I’ve also stopped using MISLDashMediaSource, as I no longer need to
replace ChunkSampleStream with MISLChunkSampleStream to access some of
the info (I’m getting it from the AlgorithmListener instead).