Move DATA_NOT_AVAILABLE and Add dataIsAvailable()
DATA_NOT_AVAILABLE used to be in the AlgorithmListener interface, but
since it’s specifically relevant to the AlgorithmInfoProvider
interface, and should be accessible to any classes implementing it,
I’ve moved it to here.

I’ve also added a method dataIsAvailable(), so that algorithms can
check if there’s no data available.