Add Average Class
This class will be used for averages that need to be calculated by
AdaptationAlgorithms. To use it, you create an instance of the class,
and pass it the type of average you want and the values to average.

I’ve also made an enum called Pythagorean with implementations of
Arithmetic, Harmonic, and Geometric averages.