Match Indices with averageWindow
The value of averageWindow should be the *number* of downloaded chunks
to include in the calculation, whereas the indices were off that by
one, leading it to interpreted as off by one.

By changing the conditions and using the number of chunks rather than
the chunk index, this is fixed.