Store and Expose Past Chunk Information
Previously, DefaultAlgorithmListener only stored information from the
most recent chunk – I have changed it to store information from all
previous chunks.

The chunk information is stored in an ArrayList, ordered with the most
recent at the end. The problem with this is there are now two sets of
chunk indices – those of the chunks relative to the stream, and those
relative to this list – and these are exposed as part of the interface.

This should be fixed at some point, by isolating clients from the
internal indices, but still allowing access to (e.g.) the 5 most
recently downloaded chunks.

Would it be worth moving the averaging functionality into the
AlgorithmInfoProviders altogether?