Include transformationFunction in AverageType
For running the harmonic mean calculation on an array of 1 element,
since the cumulativeFunction isn’t being called, the result was 1/x
instead of x.

I’ve fixed this by splitting the averages into 3 functions instead of
two – the transformationFunction is applied to each input value before
it reaches the accumulationFunction.