Write TransitionalAlgorithmListener
This listener should be compatible with the algorithms the way they
currently exist, but will get its data as a ChunkListener and
TransferListener, rather than as an AdaptiveMediaSourceEventListener.

I will then gradually convert this to match the structure I’ve
designed, which will allow me to test as I go rather than only be able
to test at the end.