Replace DashMediaSourceListener with TransitionalAlgorithmListener
This new listener is compatible with the existing code, but gets its
information from MISLDashChunkSource (an alternative implementation of
DashChunkSource) and through being registered as a TransferListener on
DashMediaSource.

Once I have tested this, I can continue to migrate all of the code
towards the code I have written, testing as I go.