Use tag variable in algorithm logging
Previously, log statements in the algorithm classes had individual
strings for each log message. This made it difficult to tell when
looking at the editor where specific log messages were coming from, as
well as what they meant.

I have changed the logs to use the name (or a portion of the name) of
the containing class, so messages can be easily identified and located.

Some of the messages still aren’t the clearest, especially as some were
using the tag to clarify the meaning and have now had that
clarification removed. I’ll address this at some point in the future by
making the messages more descriptive.