Add stall duration calculations
The TransitionalAlgorithmListener now listens to the player for stalls
and records how long the player stalls for.

At the moment, the total stall duration is all that’s recorded, and
this includes the buffering at the beginning. It might be useful to
track the stalls individually rather than just record the total stall
duration, and in that case it would be easier to ignore the initial
stall.