Allow Elastic algorithm to be selected
Added a new radio button to allow selection of the Elastic algorithm in
the player.

At the moment, the algorithm-choosing activity passes the text of the
button as a string to the main activity to indicate selection. I think
it would be better to use the ID of the button and a set of enum
constants, which would allow the text of the button to be changed
freely (and not limit us to specific casing), as well as the ID.