Update ElasticTrackSelection
I updated the references to DashMediaSourceListener to point to
TransitionalAlgorithmListener, as I’m using that class now.

I also added default value constants for the algorithm parameters, and
a new constructor for the factory which allows them to be used.

Finally, I removed some logging statements and added units to the ones
that are left.