Correct elasticTargetQueue calculation
In the C++ algorithm, elasticTargetQueue is calculated by multiplying
the max number of segments in the buffer by the current segment
duration to give the max amount of time in the buffer. In ExoPlayer,
the max buffer value is in seconds already, so there’s no need to
multiply it by the segment duration.