Create average and variance classes
Rather than use functions for these, I’ve replaced them with a class
for each kind. This avoids having to switch on an int or enum.

It also allows lazy evaluation if I want to pass the average/variance
around without evaluating it until a certain point. I can always cache
the result of the average/variance after the first evaluation if I’m
going to be evaluating it multiple times.