Move TrackGroup-related methods to AlgorithmTrackSelection
lowestBitrate(), highestBitrate(), and getNearestBitrateIndex() can all
be implemented using the information already available to a subclass of
BaseTrackSelection (i.e. the `group` and `tracks` fields and associated
methods).

Since these methods are used in several of the algorithms, it makes
sense to have them in the superclass and reuse the behaviour.

However, these changes appear to have changed how a couple of the
algorithms operate.

The BBA2 algorithm was previously jumping between the highest and
lowest quality streams (which is probably not what it was meant to do
anyway), and now jumps between the highest and third-highest streams.

The OscarH algorithm now moves up in quality much quicker than it used
to, getting to the highest quality in the space of about 7 segments,
where before it took more than 1 min 30 s.

I’m going to try and figure out what caused these changes, but first
I’m committing the code changes so that I can compare more easily.