Change minimum buffer length from 15s to 26s
As the BBA2 algorithm picks quality levels based on how full the buffer
is, allowing the buffer to shrink all the way down to 15s makes BBA2
pick a lower quality than it should when loading restarts.

Increasing the minimum buffer level to 26s (where the maximum is 30s –
allowing only a 1 segment drop) solves this problem.

All algorithms now seem to be functioning fine.