Move getThroughputSamples and getWindowSize to TransitionalAlgorithmListener
As these two functions are called by a number of different algorithms,
I’ve moved them to the listener. I chose the listener over the
superclass because the functions use information from the listener, and
so if moved to the listener only need to be called with one parameter,
rather than 2 or 3.

I’ve also changed getThroughputSamples() to create the array it’s
returning, rather than take it as an argument, as every algorithm was
just passing it an empty array.

Moving these functions to the listener should also make it easier to
change the listener’s implementation of storing past segment data.