Move estimation mode constants back to OscarHTrackSelection
As the estimation mode used to be one of the parameters for
getWindowSize and getThroughputSamples, I moved the constants to the
superclass when looking at moving those functions there.

However, since I’ve changed those functions to not rely on these
constants, and these constants are only used in OscarHTrackSeleciton,
I’m moving them back.