Remove LoadControl from algorithms
I’ve moved the getMaxBufferUs method (and changed it to milliseconds)
to the algorithm listener.

Having a consistent object which provides all of the information to the
algorithms will make them easier to understand, and this removes
reference to a static variable on the PlayerActivity class.

At the moment, the TransitionalAlgorithmListener still gets this
information by querying the PlayerActivity (though now through an
instance method), which I’m not too happy with. It should be possible
to change this later easily enough.