Remove getBufferLevel() from LogSegment
As the current buffer level is passed to updatedSelectedTrack for each
TrackSelection, the buffer level doesn’t need to be delivered by the
algorithm listener / LogSegment at all.

I’ve removed the method from LogSegment, and changed all calls to that
in the algorithms to references to the passed-in value.