Move getMpdDuration to TransitionalAlgorithmListener
Previously, this method was in two places – on the player’s video
source, which was stored as a static variable in the PlayerActivity and
accessed directly, and in MISLDashChunkSource.Factory, where it was
also accessed through a static variable in the PlayerActivity.

I have moved it to the algorithm listener, so that all information is
coming to the algorithms from the same place.