Make player object private
Previously, the player object was held in a public static field, so
that the buffer level could be calculated by algorithms referring
directly to the object.

Since the buffer level doesn’t need to be calculated separately from
the buffer level estimate passed to the algorithms (and as Android
Studio was warning that this was a memory leak), I have changed the
player field to a private instance variable.