Make logSegment private and delegate methods to it
Previously, the algorithms would get the LogSegment from the listener,
and then call methods against it to get the data. Now, I’ve changed
this so that the algorithms all get their data directly from the
listener, which accesses the LogSegment itself.

Isolating the LogSegment from access like this should make it easier
for me to replace it.