Tidy up TransitionalAlgorithmListener
I forgot to commit changes partway through, so I’ve taken a couple of
steps in one here.

I’ve removed the field logSegment, as nothing external now depends on
it, and I can get the information from the last element in the
ArrayList, rather than having to store it in a separate variable as
well as in the ArrayList. If this slows down the access too much, I can
always reinstate the field.

I’ve also renamed all the data-retrieving methods. I’ve chosen the word
“chunk” in favour of the word “segment”, for consistency with the word
use in the ExoPlayer library. I’ve also included the units of all
quantities in the method names, to make it easier to tell what units
different values are in. This should make it easier to use these values
in the algorithms and to debug them, and to change the units in future.

Finally, I’ve renamed the LogSegment class ChunkInformation, to use
“chunk” instead of “segment” as described above, and to make the
purpose of the class clearer (I think the logging functionality of the
class would be better represented in the name of the method that writes
the logs to file).