Move ChunkInformation class to TransitionalAlgorithmListener
Since I’m aiming to isolate the storing of information on downloaded
chunks from the algorithms themselves, it makes sense to make this a
nested class within the algorithm listener.

I’ve also moved the log-writing function outside of this class into the
outer class (TransitionalAlgorithmListener), as I think this makes more
sense – this way it can access the array of chunk samples without it
needing to be passed in as a parameter.