Merge branch 'fix-writing-logs-to-file'
I haven’t really fixed the problem, but there is a temporary workaround.

The problem is that the app needs to request permissions to write to
the external storage when we’re above API level 22. I tried requesting
the permission if it wasn’t already granted, but no permission dialog
appeared.

The temporary workaround is to enable the permissions for the app in
settings.