Tidy up ElasticTrackSelection
I’m starting with ElasticTrackSelection for moving the sample
processing to the listener, as it’s not very complicated and I
understand it well.

To start, I’ve removed unnecessary clutter: the comment explaining the
variables used in the C++ version of the algorithm, the unused
variables, and the early return inherited from the C++ code.

The early return should be redundant, as the
algorithmListener.chunkDataNotAvailable() check should guarantee that
the second check fails.