Update getWindowSize() and getThroughputSamples
Previously, getWindowSize was using the index of the last chunk to
calculate the window size. However, if there is a seek to the start of
the video from later on, there will be data available from later chunks
that have been downloaded, but the index of the last chunk will now be
1 (and so only the last chunk will be considered).

Changing this to the size of the array downloaded chunks should fix
this problem.

getThroughputSamples() also now calls getWindowSize() with the window
that it’s given, so it won’t request a bad index from the ArrayList if
given one.