Merge adaptiveAlgorithm and idealQuality()
Up to now, adaptiveAlgorithm() was using a counter variable to select
the lowest-bitrate index the first two times it was called. I think
this was in order to ignore the first few chunks, which are not media
chunks.

As the check algorithmListener.chunkDataNotAvailable() (in
idealQuality()) should cover this case, I have replaced the two methods
with one called adaptiveAlgorithm().