Replace ternary operators with min() and max()
The ternary operator is used in several places to select the smaller or
larger of two values, or to clip a value to a certain minimum or
maximum value (e.g. not letting the suggested index go below the
minimum or above the maximum).

In both cases, I think it’s clearer to use Math.min() and Math.max().
Also, I think that won’t evaluate it’s arguments more than once,
whereas I think the ternary will evaluate the condition and the
selected output, which may be the same statement.