Use lastChunkIndex() in getThroughputSamples()
getThroughputSamples() was using the chunk index of the last chunk to
calculate chunk indices for retrieving chunks from the ArrayList. As
before with getWindowSize(), this will break when seeking occurs, as
chunks are not stored according to their own index, but instead are
stored sequentially as they finish downloading.

Changing this to lastChunkIndex() (which uses the size of the array to
find the index of its last elements) should fix this.