Replace normalised sample calculation and maximum sample calculation
I’ve replaced the calculation of normalised throughput samples within
OscarTrackSelection with a call to a new method within the listener.

Note that the maximum throughput sample is also calculated within
getNormalisedSamples(), which is a bit inefficient. However, moving
things to separate methods within the algorithm listener gives us more
flexibility, which I think is worth it.

Also, it should be possible to cache these results within the listener
as they are calculated, which would solve this problem.