Move oscarKumarParEstimation() to listener and revert changes
I was gradually replacing the calculations in this method with calls to
smaller methods in the listener class. However, I’m not sure if that’s
worthwhile, as I don’t know how much the different parts are reused,
and I was making the calculations less efficient.

I have decided to undo a lot of the changes that I’ve made, and move
the whole method to the listener as-is. The changes I had made can be
found at the commit before this one, as well as in the branch
‘change-oscarKumarParEstimation’.

As I move the other sample processing to the listener, it should become
more clear which bits of this method should be split out, at which
point I can reintroduce the changes. One potential advantage of the
changes is that they’ll make the code more understandable, which might
be desired.

I have also moved the helper method computeS() to the listener along
with oscarKumarParEstimation().