Rename counter and remove unnecessary declaration
The counter was called `ii`, presumably to prevent conflict with
another counter called `i`. As there is no conflict here, I’ve renamed
it `i`, since the name is more standard.

I’ve also moved the declaration to the initialisation section of the
for loop, as that puts it closer to where it’s used.