Use separate counter variables for each loop
Rather than use one counter, resetting the value at each for loop, I’ve
changed the code to create a new counter (with the same name) in each
for loop.

From what I’ve read, there is not a major performance hit, and this
allows the loops to be split up more easily (don’t have to add a new
variable declaration at that point).

I’m also more used to seeing code with the declaration in the loop
header.