Remove window from exponential average and variance classes
These classes used to work by being given an array of value and a
window parameter, which indicated how many of those values should be
used.

I’ve removed the window parameter, so that the classes just give the
average and variance of the array of values you give them. The window
parameter is not needed anymore, as getThroughputSamples() can be
called to get just the needed samples.