Replace sample arrays with lists
By depending on the List interface rather than an array of primitives,
I reckon we get more flexibility. Also, we can use the List methods and
those defined in the Collections class.

Using Lists may be slower than using arrays of primitives. I’ve read
that it’s not much slower, and that the performance difference
shouldn’t be relevant.