Pass max buffer level to listener constructor
Previously, the player activity was passing itself to the listener, and
the listener was then calling a method on the activity to get the value
for the max buffer length.

This seems overwrought, especially as the max buffer length so far
isn’t changing dynamically. I’ve changed it to pass the buffer length
to the listener when it’s created – we can update the value the
listener has dynamically if we ever need to.

I’ve also changed the listener from being a static variable to an
instance variable. I don’t think there’s any need for the listener to
be a static variable, and I think this makes more sense.