Remove inc flag and move chunkDataNotAvailable() check
Previously, a flag was being used to ignore the first two chunks that
were downloaded, as they were null. That is now encapsulated inside
chunkDataNotAvailable(), so there’s no need to have those checks.

There’s also no need for the methods adaptiveAlgorithm() or
idealQuality(), so I’ve removed them. In each algorithm,
updateSelectedTrack() now calls the method that carries out the
algorithm.

I also moved the chunkDataNotAvailable() check outside of the algorithm
method in each case, as it’s not really part of the adaptation
algorithm, and makes more sense separated.

I changed the way the DASHTrackSelection is making sure it has values
as well. It used to be more complicated, involving a flag and checking
whether the estimate from the bandwidth meter was available. I’ve
changed it so it just checks whether the bandwidth estimate is
available – if it is, it uses it in the algorithm, otherwise it selects
the lowest index.

I also replaced the idealQuality() method here with
findBestRateIndex(), which does the same thing and is implemented in
the superclass.

Note that the DASHTrackSelection does not currently work – the
bandwidth meter it depends upon is not being registered as a transfer
listener on the media source, and so receives no data it can use to
make bandwidth estimates. I will fix this soon.