Replace deprecated newSimpleInstance() call
The version of ExoPlayerFactory.newSimpleInstance() that was being used
is deprecated, so I’ve replaced it with the suggested alternative,
which takes a RenderersFactory rather than a Context.

I’ve gone with DefaultRenderersFactory, constructed with as many
defaults as possible. I don’t think it’s going to make any difference
to us for now.