Fix PlayerActivity
I think there were a couple of things I missed in resolving the merge
conflict, but I’m fixing it now.

I’ve recreated the algorithmType field, which got lost along the way,
and changed the DefaultLoadcontrol to a MISLLoadControl.

I’ve also updated the switch in chooseAlgorithm() to use the algorithm
type cases instead of the string names of the algorithms.