Switch on button ids for choosing algorithm
I had been thinking that when switching on the button IDs, I’d be
switching on a string value, but it turns out those IDs are int enum
constants. So instead I’m switching on an int and matching against
those values.

I had also forgotten to pass the ID to the switching method rather than
the buttons’ text, so I’ve fixed that now.