Write AdaptationAlgorithm implementations
The implementations create the required track selection factories and
listeners for each adaptation algorithm.

The idea is that this would be cleaner than having separate cases for
(e.g.) whether to create a BandwidthMeter or a
TransitionalAlgorithmListener (DashTrackSelection and
AdaptiveTrackSelection vs. the others). However, I’ve had to write a
fair bit of quite similar code, and I’m not sure if I’ve gained a
massive amount.

On the plus side, this has made it easy to fix the DashMediaSelection,
as it’s now straightforward for one algorithm to use a BandwidthMeter.