Incorporate AdaptationAlgorithms
The player activity now has reference to an AdaptationAlgorithm, from
which it gets a ChunkListener, TransferListener,
ExoPlayer.EventListener, and TrackSelection.Factory.

I’ve removed the basic exoplayer option, as I don’t think we need it.
It should be easy to reimplement if needed.

I’ve also added null checks for the listeners, as they may not always
be needed, and we want to avoid null pointer errors.

I removed the getInfosVideo() method because it was interacting
directly with the algorithm listener, and I don’t think it was doing
anything (certainly nothing was visible on the screen). We can get it
back easily enough if we need to.