Create size-based sampler
The size-based sampler will sample the throughput every x bytes.

One issue with this is that, because of the way the
onBytesTransferred() callback works, I can’t guarantee it will be
exactly every x bytes that a sample is made, only that it won’t happen
more often than every x bytes.

In practice, I think this should be fine, but it would be nice to be
more precise. Then again, the TransferListener probably works this way
because of limits in the underlying protocol – it’s chunks/segments at
some point.