Write sample processor interface
This interface covers the methods the adaptation algorithms will call
to get the information they need.

I’ve gone with separate methods for getting the (e.g.) size of the last
chunk and the size of the last throughput sample – for chunk-based
sampling these will be the same, but keeping them separate makes more
sense in general and allows for separately tracking throughput samples
and chunk throughput samples.