Rewrite algorithm listener as sample processor and update elastic algorithm
I have rewritten the listener as a SampleProcessor, and offloaded
sample collection to another object (this can be different for each
algorithm – for the elastic algorithm I’ve used a ChunkBasedSampler).

Note that at the moment this breaks writing the logs to file – a lot of
the work for that was done in the toString() method of the
ChunkInformation class, which I have gotten rid of here now that I’m
separating throughput samples from chunks. I will need to fix this at
some point.

I still haven’t sorted out the mpd duration or stall duration, either.

I’ve also updated the elastic algorithm to use the new structure – it
uses a ChunkBasedSampler and a DefaultSampleProcessor.