Include mpd duration in default sample processor
I have put the mpd collection in the ChunkStore interface, though I’m
not happy about this long-term. As it’s not really to do with
individual chunks, I think it should be in a separate interface, or
possibly reported directly to the sample processor.

Providing the mpd duration again fixes a runtime error I was getting in
the OscarH algorithm, where an mpd duration of 0 caused the algorithm
to attempt to create an array of size -1.