Replace toString() method for logging chunk info
Previously, the toString() method on the ChunkInformation (originally
LogSegment) class was used to provide the lines for the log. Since that
class is gone, the log writing has been broken, and here I’ve partially
fixed it.

The sample processor doesn’t currently have access to some of the
information: the arrival time of the chunks, the stall duration of the
player when each chunk was downloaded, the actual data rate of each
chunk, and the buffer level when each chunk was downloaded. So at the
moment it’s writing zeroes into the file for those values.

I think I need to move the logging into a separate class – either a new
one or into the ChunkStore implementations.