Add stall and buffer methods to ChunkStore
I’ve decided to keep the logging in the chunk store for the moment,
which means the interface needs a way to be notified of stalls and of
updates to the buffer estimate.

To get these things, the chunk store will need to be passed to the
exoplayer event listener and also to the track selection factory.