Bring back LogSegment
I’ve renamed the class LogEntry and given the fields better names
(including putting the units in the names for clarity).

At the moment, the current buffer length and the current stall duration
are not updated in the chunk store, and so aren’t logged. This should
be straightforward to fix.

It would be good to update the toString() method in the LogEntry class
to use String.format() instead of the while loops that are currently
there, or possibly use an XML format for logging instead of plaintext.