Update buffer level from arbiter track selection
I’ve renamed the algorithmListener parameter sampleProcessor, to
reflect the changed interface. I’ve also added a second parameter for
the chunk store, which is where the track selection sends the updated
buffer estimate before updating the selected track.

The logging for the buffer level is now working as expected for this
algorithm, and showing sensible values.