Remove algorithm object
As it was, a new class would have to be created for every combination
of components we wanted to use (e.g. to use one algorithm with a
different kind of sampling). That leads to a lot of duplicated code and
extra classes.

Also, when this software is used it’s going to be automated, and each
component should be individually configurable at runtime.

I’ve rewritten the chooseAlgorithm() function to switch over the enum
constants selected when running the player, and do the configuration
accordingly. I’ve also renamed it configureRun(), to reflect what it
does a bit more clearly.