Prevent infinite loops in LogEntry class
The while loops that were being used to pad the column strings with
spaces used a != condition, which meant that they caused infinite loops
if the value started wider than the padding limit.

In onPause(), this was causing the app to hang after the back button
was pressed from the player. When moved to onStop(), I think the
process was killed while in the infinite loop.

I have replaced these while loops with a template string and
String.format(). It would probably be nice to be able to pass in the
column sizes rather than hardcoding them – I might come back to that.