Remove logging code from DefaultSampleProcessor
The sample processor was storing an array of chunks and an array of log
entries.

Now that it doesn’t need to do any logging (and I know that the
algorithms don’t need any information on previous chunks – just the
throughput samples), I’ve gotten rid of the log array and replaced the
chunk array with just a reference to the most recently downloaded chunk.