Move logging code to DefaultChunkStore
Rather than collect chunks from the chunk-based sampler, the logging
code will now collect data from the media source, using the
AdaptiveMediaSourceEventListener.

As this is just for logging, the timing of the information delivery is
not crucial, so the AdaptiveMediaSourceEventListener should be fine.
One possible issue is the difference in the load duration values that
this reports – I’ll need to look at these to see how significant the
difference is. If needed, we can pass that to the DefaultChunkStore
from the chunk listener.

As the chunk index is not available in the
AdaptiveMediaSourceEventListener, I’ve hardcoded a calculation based on
the media start time. However, as this depends on a constant segment
duration, I think a better idea would be to use only the media start
time and not the chunk index at all.