Calculate stall duration by extending ExoPlayer.EventListener
I’ve removed newStall() from the ChunkStore interface, and changed it
to implement ExoPlayer.EventListener.

DefaultChunkStore uses the onPlayerStateChanged() callback to start a
clock if the player was playing and is now buffering. This will also
count seek operations, but we’re assuming for the moment that there
won’t be any during the run. If needed, we can use
onPositionDiscontinuity() to detect those.

I haven’t tested this code properly yet, as I have no way of
artificially creating stalls. If I move away from the wifi far enough
that it disconnects, the player won’t restart when I reconnect.