Move giveMpdDuration() call out of the factory
Up to now, the manifest duration was given to the chunk listener in the
factory’s createDashChunkSource() method. I think I put it here
originally with the idea that it should happen as early as possible,
but I think it makes more sense in the MISLDashChunkSource constructor.

For one thing, it will only be happening slightly later, so I don’t
think the time difference will have any effect. For another, I think it
makes more sense to keep things like this out of the factory, as the
factory’s only concern should be to create the MISLDashChunkSource
instance.