Move logger buffer updates to MISLDashChunkSource
I had started passing these buffer updates from the individual track
selections to the chunk store (logger), with the idea that the
calculation was already done for me there, so I should make use of that.

However, since we’re using multiple track selection classes, we’d have
to have that code duplicated in all of them. Instead, I’m doing it from
MISLDashChunkSource, which will mean we’ll only have to have it in one
place.

The calculation I have duplicated from inside the DashChunkSource. This
does mean that it’s calculated identically in two places, but I
couldn’t see any way around this.

One issue with this is that the values are passed to the chunk store on
the thread responsible for updating the track selections (I think this
is the execution thread?). It might be better to move this to a
different thread, like with the AdaptiveMediaSourceEventListener, so
it’s not slowing down the operation of the player.