Store chunk logs by the chunk index
Originally, chunks were stored by their index (in the LogSegment
class). I hadn’t realised this was important, and so changed it to just
store them in download order (at the time they were used for throughput
sampling, and I felt it made more sense to put the samples in download
order).

However, since they’re for logging (and not for providing throughput
samples), it makes sense to have them ordered by chunk index, and it
will better facilitate chunk replacement.

As the chunk index calculation assumes constant-duration chunks though,
this won’t work if chunk duration is variable within the video. It
might be better to store chunks by their start or end time to account
for this, but I’m not sure how this would work if one 4 second chunk
was replaced with one 8 second chunk, for example.