Use flag for stalls
As the player enters the buffering state at the beginning, before
playback begins, and then enters the ready state once playback begins,
checking for this state transition was causing the end of a stall to be
recorded – as the stall clock hadn’t been started (and so was at 0),
the stall duration was being updated to the current value of
SystemClock.elapsedRealtime().

To fix this, I have introduced a currentlyStalling flag, and now check
for that when detecting the end of a stall.