Move info from javadoc param to description
I think the description of the behaviour of the sampler makes more
sense in the javadoc body than in the parameter description, as it’s
functionality of the sampler, not of the parameter.

I’ve also renamed a couple of variables to indicate the correct units
(sampleThresholdBits was actually holding a value in bytes).