Stop size-based sampler clock between transfers
The sampler was counting time between transfers as time spent
downloading, which was especially bad when loading was paused because
the buffer was nearly full.

I have changed it to track starts and ends of transfers, and only count
time spent transferring for throughput samples.

I have also restructured the code with internal methods that make the
sampling process easier to think about.

There is still the problem of samples being split across periods of
time where loading stopped – these samples will not reflect network
changes while loading is not happening.

The suggested solution is to make a premature sample when loading
stops, and begin a new sample when loading starts again. To do this, I
will need some way of detecting that the player has stopped loading –
this can probably be done in conjunction with the load control.

Another, simpler solution would be to always deliver a premature sample
when a transfer ends. For larger transfers, the size threshold would
dominate, and for smaller transfers the transfer rate would dominate.