Use size-based estimation
I have added a button to the algorithm chooser that allows you to
choose the basic algorithm using size-based estimation.

It works, and as expected, smaller thresholds give more variation in
the rate than larger thresholds.

I changed the text on the buttons to be in lowercase, as well, because
I think it looks nicer.

I have also deleted the TransitionalAlgorithmListener class, as it’s no
longer needed, and the other SampleProcessor interface, as I’m not
using it and it was confusing having two of them. I think I have
deleted these before – they must have come back in a merge at some
point.