Implement ExoPlayer.EventListener to detect loading stop
By listening to the player and using the onLoadingChanged() method, the
sampler can deliver a premature sample if there’s going to be a delay
before loading continues.

I’ve also restructured the sampler to treat the time elapsed as it does
the bytes transferred – update the sample clock and the sample duration
every time onBytesTransferred() is called.

When structured this way, the pauseSampling() method is no longer
needed, and nothing needs to be done at the end of a transfer.

I’ve also renamed some fields and methods:

* byteClock -> sampleBytesTransferred

    - this is clearer, as it corresponds better with the time
measurements

* sampleStartMs -> sampleClockMs

    - as the value is updated as the sample continues, it doesn’t
always refer to the sample start time

* startedSampling() -> currentlySampling()

    - I think this name says what it means more clearly