Prevent samples smaller than 1 ms from being made
Sometimes a premature sample is shorter than 1 ms in duration, which
means that the measured duration in ms is 0 – this gives a throughput
estimate of infinity.

If we’re happy to target higher APIs, we could use
elapsedRealtimeNanos() instead of the standard (which is in ms), and
that should solve the problem as well.

It might also be good to allow the minimum threshold to be specified
when the sampler is created.